requirements.txt
langchain
langchain-community
langchain-core
faiss-cpu
groq
huggingface-hub
sentence-transformers
pypdf
PyMuPDF
tqdm
python-dotenv
langchain-groq
python-dotenv

langchain
langchain-groq
PyMuPDF
python-dotenv
faiss-cpu
